requirements.txt
# AI API
anthropic==0.76.0

# Web scraping
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0

# Email
python-dotenv==1.0.1

# Data processing
python-dateutil==2.8.2

# Optional: pro pokročilejší scraping
# playwright==1.45.0

# Web frontend
flask==3.0.0
flask-limiter>=3.5.0

# Retry logic
tenacity>=8.2.0
